.ipynb_checkpoints/requirements-api-checkpoint.txt
torch==2.0.1
torchvision==0.15.2
flask==2.3.2
pillow==10.0.0
numpy==1.24.4
gradio
requests